requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6
boto3==1.34.34
python-dotenv==1.0.1
requests==2.31.0
